Deleting games can only be done by the creator

+++ RecreateMe/RecreateMe/Scheduling/Handlers/DeleteTeamGameRequestHandler.cs
@@ -11,6 +11,11 @@
 
         public DeleteTeamGameResponse Handle(DeleteTeamGameRequest request)
         {
+            var game = _teamGameRepository.GetTeamGameById(request.GameId);
+
+            if (game.Creator != request.ProfileId)
+                return new DeleteTeamGameResponse {Status = ResponseCodes.NotCreator};
+
             _teamGameRepository.DeleteGame(request.GameId);
 
             return new DeleteTeamGameResponse();
@@ -20,9 +25,12 @@
     public class DeleteTeamGameRequest
     {
         public string GameId { get; set; }
+
+        public string ProfileId { get; set; }
     }
 
     public class DeleteTeamGameResponse
     {
+        public ResponseCodes Status { get; set; }
     }
 }
